Add a pkg-config file.
This allows other libraries on top of librevisa and also librevisa-using
applications to easily integrate it into various build systems on
various architectures and OSes, both for normal compiles and cross-compiles.

The librevisa configure script knows which additional libraries are used
or not (e.g. libusb-1.0) and correctly places that info into the
librevisa.pc file for use by pkg-config.

The setup can be tested as outlined below. You can use PKG_CONFIG_PATH
if librevisa is not installed in a standard path like /usr or /usr/local.

Example:
  $ ./configure --prefix=/tmp/testing && make install
  $ PKG_CONFIG_PATH=/tmp/testing/lib/pkgconfig pkg-config ...

Using pkg-config manually:
--------------------------

$ pkg-config --modversion librevisa
0.0.20121230

When compiled without libusb-1.0:
---------------------------------

$ pkg-config --cflags librevisa
-I/usr/local/include

$ pkg-config --libs librevisa
-L/usr/local/lib -lvisa

$ pkg-config --libs --static librevisa
-L/usr/local/lib -lvisa

With libusb-1.0:
----------------

$ pkg-config --cflags librevisa
-I/usr/local/include -I/usr/include/libusb-1.0

$ pkg-config --libs librevisa
-L/usr/local/lib -lvisa

$ pkg-config --libs --static librevisa
-pthread -L/usr/local/lib -lvisa -lusb-1.0 -lrt