Use array during global init
The resource creators are registered during global init, at which point the
list may be uninitialized which leads to errors during initialization
and/or creators being forgotten when the ctor is run.

Using an array and assuming proper NULL initialization solves the problem
for most platforms. It does not work when NULL is not represented by an
all-zeros pattern.